[torod, torodb] Created subproject 'backend', to factor all code related to persistence of data
The backend (interface Backend and abstract class AbstractDbBackend) abstracts all the logic related to persistence storage initialization and management. In the case of a database-based backend, it creates and manages the connection pool and the database-dependent initialization code, and provides it to the executors via injection.

This new backend fixes some abstraction problems in the previous code, where database initializacion and driver configuration were handled by upper layers. Also improves configuration of the connection pool and improving on the documentation. Also, for PostgreSQL-based backends, a diversion is stablished to easily switch, in the future, without adding now extra compile-time dependencies, to the pgjdcb-ng driver, instead of the "official" driver. The application name is set, too, so that connections may be distinguished from a PostgreSQL server by querying by application_name on pg_stat_activity.

Code should be cleaner, better abstracted and more flexible with this commit.