Created a MongoDB command to create path views
The new command "createPathViews" has been created. It has the following
syntax: {createPathViews: <collectionName>} and, when executed, it
creates one SQL view on the backend database for each possible document
path on the backend. More precisely, the only created views are the ones
that start on the root of each document and do not include arrays