Moved classes not driver dependent to db-wrappers
Classes AbstractDbBackend and DbBackendConfiguration were in the driver/
subpackage but are not related to the driver. Moved to parent.

Added an abstract protected method to AbstractDbBackend to force
implementors to define how transaction isolation is performed for every
DbBackend, as some of them may be run under different isolation modes
(to still achieve the same effect).