Added a command to execute SQL selects
A new ToroDB dependient command has been added. It is called
_sql-select_ and can be used to... execute selects on the backend
database. We know this is a security hole and the command will change
in future to be able to only select data from ToroDB related tables.
Meanwhile, the transaction where the command is executed is forced to
be read only.

The syntax of the command is the following:
{"sql-select": <collection name>, query: <standard sql query>}

For example, executing the following on mongo console (using the 
correct db!) will return all content from torodb.collections table
(the table where ToroDB stores collection metadata)
db.test.runCommand({
    "sql-select": "test4", 
    "query": "select * from torodb.collections"
})

The output is a cursor that contains one document for each row returned
by the query. Right know, the cursor only contains one batch. That means
all data returned by the query will be stored on memory and returned on
a (potentially) big response. In future a real cursor will be returned,
so it could be consumed with getMore.

Right know the collection provided as "sql-select" input field is not
really important and it is only used on the cursor namespace. In future
it could (or could not) constraint the schema where the query must be
executed.