Integration test configurations for travis
Two enviroment variable control configuration and tests 
that are ignored or not:

* com.torodb.integration.config.Protocol: Specify wich 
protocol to be used.
* com.torodb.integration.config.Backend: Specify wich 
backend to be used.

Depending on the backend and protocol test are ignored and 
ToroDB is configured accordingly to use that protocol or that 
backend.
We use environment variables since travis can be configured 
to start a battery of test with all configuration combinations 
allowing to test all ToroDB configurations in separate jobs 
and with different services.
Jstests are now wrapped to an enum and can be configured, 
in block or separately, to be in category working, failing, 
etc. using annotations on enum value. Junit rule are used to 
permit ignore or not a jstest based on those annotations and 
the current configuration.