isMaster command bypass the used database check
Since MongoWP command API changed, there is a check that abort any
command execution on a non supported* non special** database. This has
sense in almost all commands, but there are some of them on which this
constraint is too restrictive. There is now a white list of commands
that bypass the given constraint so they can be executed on any
database. At this moment, isMaster is the only command on that list, but
new commands can be added if needed.

*: Right now, ToroDB only support a single database that whose name is
configured at startup
**: Special databases are _admin_ and _local_