Changed min wire version to 0
A previous commit modify the value of minWireVersion returned on command
ToroDB, changed it from 0 to 3. It has sense because our support of the
new API is better than the older (2, specially MongoDB 2.4). But it
seems that there are drivers/clients that do not correctly interpeter
that value as we expect. As an example, MongoDB Console misunderstood
(or at least, does not understan the value as we expected) the value and
it uses the older API when minWireVersion is != 0.

This commit reverts the modification and set it to 0 again so at least
the official console will use the new api by default.