Improved the efficient of SubDocHelper
SubDocHelper.java is used to escape fields whose names are '_*index'
or '_*did' with an extra '_'. This is done to avoid SQL errors when a
document contains an attribute whose name is one of the reserved column
names that all subdoc tables have (did and index).

That proces was done using regexp, but (as expected) it was quite slow,
specially when a lot of documents were inserted in batchs. An
anternative implementation has been done. It is quite verbose but it
should be very efficient.

A unit test has been created to check the implementation.