Improved SplitDocument and its builder to do not copy memory
ToroDB creates SplitDocument instances using its Builder. That Builder
creates immutable copies of its structures to be reusable (= the same
builder can be used to create another SplitDocument). This should be
the most efficient way to create a bunch of SplitDocuments, but by
construction, ToroDB needs to create a new Builder each time a
SplitDocument has to be created, so it was a waste of resources. As
it is more difficult to change the way ToroDB creates SplitDocuments
than modify the way the builder works, the second option was choosen.