Fixed a couple of bugs related to specific configs on integration tests
Some configuration options were ignored on the integration tests. For
example, the default database was drop and created, even if a different
database is provided on the config file.

At the same time, a new log sentence have been added. It informs wich
js test is being running, so it will be easy to which test is failing
just looking at the logs.