Some improvements to the integration test framework.
The integration test framework had a couple of bugs and read it was more
difficult than needed. It still uses parametrized tests to lauch an 
execution per js file and filters them by environment variables (in fact
the interface/configuration didn't change) but it has some differencies:

* A new class called IntegrationTestEnvironment is the one that stores
the combination between protocol (single mongo node or mongo on
replication) and backend (postgres or greenplum).
* Each IntegrationTestEnvironment combination has its own test defined.
* On a given IntegrationTestEnvironment, a test is classified as:
** WORKING: if it must succeed
** FAILING: if it must fail
** CATASTROPHIC: if it could freeze ToroDB or the integration framework.
This kind of scripts represents serious bugs on ToroDB that might be
fixed ASAP.
** FALSE_POSITIVE, NOT_IMPLEMENTED or IGNORED: different kind of tests
that will be ignored
* The framework will only run tests assigned to its 
IntegrationTestEnvironment.
* Scripts classified as WORKING must succeed (or the test will fail)
* Scripts classified as FAILING must fail (or the test will fail)
* Scripts classified as CATASTROPHIC wont be executed, but a warning
will be printed.
* Scripts classified as any other classification will be ignored (the
junit test will be marked as ignored).