Changed the thread model of initial sync phase
Initial sync phase is now done concurrently. Preliminary benchmarks show
that this implementation is at least 50% faster than the older one, but
some paramenters values should be studied to know if we can improve that
number.

As new executor services have to be used now, the execution service
injectation has externalized to a specialized module and the current
implementation creates a single cached thread pool that is shared in
almost all places where threads are needed. This is a simplification and
it may be necessary to modify that on the future, as this kind of
executor service is bad when a lot of task are requested, as it creates
an unbounded number of threads.