Fixed retrier behaviour when the callable fail because a RuntimeException
Before this commit the behaviour of the retrier was not compatible with
its javadoc. Javadoc said that any runtime exception different than
RollbackException will prevent a new execution, but the implementation
ignore that and treated all runtime exceptions (but 
RetrierAbortException) as checked exceptions.

This has two implications:
	+ Users of the retrier interface can found new scenarios where the
execution is not retried
	+ When the callable throw a runtime exception (different than 
RollbackException) the exception thrown by the retrier is not
RetrierGiveUpException but the same exception the callable thrown.