Several changes on the repository structure
There project has now three main folders: engine, server and stampede.
The first one is the "core" of the ToroDB technology and the other two
are different ways to use it. Stampede is designed to replicate data
from a MongoDB replica set and store it on a SQL backend. Server is 
designed to be a almost transparent substitute of a Mongod node.

Right now there are more folders on this repository that will be
eventually move to their respective repositories or folders.