Changed the max test timeout to 10s on MvccMetainfoRepositoryTest
This test was making that some compilations fail on travis. The timeout
is there to stop compilations if a deadlock is found, but I guess this
failures are due to the slow machines Travis use. I changed the timeout
form 1s to 10ms. If these failures were due to a randome deadlock, they
should fail anyway.