Changed the folder of integration tests to the default value
Our first iteration on the integration framework used a specific folder
under each project to classify tests as unit test or integration tests.
It works fine, but IDEs require some configuration to recognized the 
folder. We have change that to use a more standard approach on which
tests whose class name ends with *Test are unit tests and those that 
ends with *IT are integration tests and therefore must be ignored by 
surefire pluging and executed with failsafe plugin

The previous commit tryed to use a JUnit Category instead of the class
name, but it has some problems related to the class being looked for
before the project is installed.